+++ ephios/event_management/forms.py
@@ -10,7 +10,7 @@ from guardian.shortcuts import assign_perm, remove_perm
 
 from ephios.event_management import signup
 from ephios.event_management.models import Event, Shift
-from ephios.helpers.widgets import CustomDateInput, CustomTimeInput
+from ephios.extra.widgets import CustomDateInput, CustomTimeInput
 from ephios.user_management.models import UserProfile
 from ephios.user_management.widgets import MultiUserProfileWidget
