+++ ephios/event_management/views.py
@@ -25,7 +25,7 @@ from guardian.shortcuts import get_objects_for_user, get_users_with_perms
 
 from ephios.event_management.forms import EventForm, ShiftForm
 from ephios.event_management.models import Event, Shift
-from ephios.helpers.permissions import get_groups_with_perms
+from ephios.extra.permissions import get_groups_with_perms
 
 
 class HomeView(LoginRequiredMixin, TemplateView):
