+++ ephios/extra/context.py
@@ -1,7 +1,7 @@
 from django.templatetags.static import static
 from django.utils.translation import get_language
 
-from ephios.helpers.signals import footer_link
+from ephios.extra.signals import footer_link
 
 
 def ephios_base_context(request):
@@ -14,4 +14,7 @@ def ephios_base_context(request):
     if get_language() == "de-de":
         datatables_translation_url = static("datatables/german.json")
 
-    return {"footer": footer, "datatables_translation_url": datatables_translation_url}
+    return {
+        "footer": footer,
+        "datatables_translation_url": datatables_translation_url,
+    }
