+++ ephios/user_management/forms.py
@@ -13,7 +13,7 @@ from django.utils.translation import gettext as _
 from django_select2.forms import Select2MultipleWidget, Select2Widget
 from guardian.shortcuts import assign_perm, remove_perm
 
-from ephios.helpers.widgets import CustomDateInput
+from ephios.extra.widgets import CustomDateInput
 from ephios.user_management.models import QualificationGrant, UserProfile
 from ephios.user_management.widgets import MultiUserProfileWidget
